requirements.txt
requests
pyyaml
flask
fastapi
openai
pydantic
uvicorn
python-dotenv
